Add extra validation to file uploads

+++ app/refresh/players/refresh.js
@@ -6,6 +6,11 @@ const { mapPlayer } = require('./map-player')
 const refresh = async (path, token) => {
   const workbook = XLSX.readFile(path)
   const worksheet = workbook.Sheets.ALL
+
+  if (!worksheet) {
+    return { success: false }
+  }
+
   const players = XLSX.utils.sheet_to_json(worksheet)
   await deleteFile(path)
   const mappedPlayers = players.map(mapPlayer)
